\name{MMSE}
\docType{data}
\alias{MMSE}
\title{Mini Mental State Examination}
\description{
  Four primary studies which report the accuracy of the Mini Mental State Examination by giving the screening outcomes against the Cambridge Mental Disorders of the Elderly Examination (CAMDEX).
}
\usage{MMSE}
\format{A vector containing 5 variables and 4 observations. The variables are:
\tabular{lll}{
  \bold{TP} \tab \code{numeric} \tab number of true positives \cr
  \bold{FN} \tab \code{numeric} \tab number of false negatives \cr
  \bold{FP} \tab \code{numeric} \tab number of false positives \cr
  \bold{TN} \tab \code{numeric} \tab number of true negatives \cr
  \bold{study} \tab \code{factor} \tab primary study name
}
}
\source{Botella et al. 2013, table 2}
\references{
  Botella, J., Huang, H., Suero, M.(2013). Multinomial tree models for assessing the status of the reference in studies of the accuracy of tools for binary classification. \emph{Frontiers in Psychology},4:694. 
  \url{http://www.frontiersin.org/Journal/Abstract.aspx?  s=956&name=quantitative_psychology_and_measurement&ART_DOI=10.3389/fpsyg.2013.00694}
}
\keyword{datasets}